--- src/language/helpers/checks.ts
@@ -1,36 +1,0 @@
-import {
-    isSdsAttribute,
-    isSdsClass,
-    isSdsEnum,
-    isSdsFunction,
-    isSdsSegment,
-    SdsArgument,
-    SdsClassMember,
-    SdsDeclaration,
-    SdsTypeArgument,
-} from '../generated/ast.js';
-
-export const isInternal = (node: SdsDeclaration): boolean => {
-    return isSdsSegment(node) && node.visibility === 'internal';
-};
-
-export const isNamedArgument = (node: SdsArgument): boolean => {
-    return Boolean(node.parameter);
-};
-
-export const isNamedTypeArgument = (node: SdsTypeArgument): boolean => {
-    return Boolean(node.typeParameter);
-};
-
-export const isStatic = (node: SdsClassMember): boolean => {
-    if (isSdsClass(node) || isSdsEnum(node)) {
-        return true;
-    } else if (isSdsAttribute(node)) {
-        return node.isStatic;
-    } else if (isSdsFunction(node)) {
-        return node.isStatic;
-    } else {
-        /* c8 ignore next 2 */
-        return false;
-    }
-};
