+++ src/language/helpers/nodeProperties.ts
@@ -1,10 +1,15 @@
 import {
     isSdsAssignment,
+    isSdsAttribute,
     isSdsBlockLambdaResult,
+    isSdsClass,
     isSdsDeclaration,
+    isSdsEnum,
+    isSdsFunction,
     isSdsModule,
     isSdsModuleMember,
     isSdsPlaceholder,
+    isSdsSegment,
     SdsAbstractCall,
     SdsAnnotatedObject,
     SdsAnnotationCall,
@@ -17,6 +22,7 @@ import {
     SdsCallable,
     SdsClass,
     SdsClassMember,
+    SdsDeclaration,
     SdsEnum,
     SdsEnumVariant,
     SdsImport,
@@ -38,6 +44,39 @@ import {
 } from '../generated/ast.js';
 import { AstNode, getContainerOfType, stream } from 'langium';
 
+// -------------------------------------------------------------------------------------------------
+// Checks
+// -------------------------------------------------------------------------------------------------
+
+export const isInternal = (node: SdsDeclaration): boolean => {
+    return isSdsSegment(node) && node.visibility === 'internal';
+};
+
+export const isNamedArgument = (node: SdsArgument): boolean => {
+    return Boolean(node.parameter);
+};
+
+export const isNamedTypeArgument = (node: SdsTypeArgument): boolean => {
+    return Boolean(node.typeParameter);
+};
+
+export const isStatic = (node: SdsClassMember): boolean => {
+    if (isSdsClass(node) || isSdsEnum(node)) {
+        return true;
+    } else if (isSdsAttribute(node)) {
+        return node.isStatic;
+    } else if (isSdsFunction(node)) {
+        return node.isStatic;
+    } else {
+        /* c8 ignore next 2 */
+        return false;
+    }
+};
+
+// -------------------------------------------------------------------------------------------------
+// Accessors for list elements
+// -------------------------------------------------------------------------------------------------
+
 export const annotationCallsOrEmpty = function (node: SdsAnnotatedObject | undefined): SdsAnnotationCall[] {
     if (!node) {
         /* c8 ignore next 2 */
@@ -51,15 +90,12 @@ export const annotationCallsOrEmpty = function (node: SdsAnnotatedObject | undef
         return node?.annotationCalls ?? [];
     }
 };
-
 export const argumentsOrEmpty = function (node: SdsAbstractCall | undefined): SdsArgument[] {
     return node?.argumentList?.arguments ?? [];
 };
-
 export const assigneesOrEmpty = function (node: SdsAssignment | undefined): SdsAssignee[] {
     return node?.assigneeList?.assignees ?? [];
 };
-
 export const blockLambdaResultsOrEmpty = function (node: SdsBlockLambda | undefined): SdsBlockLambdaResult[] {
     return stream(statementsOrEmpty(node?.body))
         .filter(isSdsAssignment)
@@ -67,42 +103,33 @@ export const blockLambdaResultsOrEmpty = function (node: SdsBlockLambda | undefi
         .filter(isSdsBlockLambdaResult)
         .toArray();
 };
-
 export const importedDeclarationsOrEmpty = function (node: SdsQualifiedImport | undefined): SdsImportedDeclaration[] {
     return node?.importedDeclarationList?.importedDeclarations ?? [];
 };
-
 export const literalsOrEmpty = function (node: SdsLiteralType | undefined): SdsLiteral[] {
     return node?.literalList?.literals ?? [];
 };
-
 export const classMembersOrEmpty = function (
     node: SdsClass | undefined,
     filterFunction: (member: SdsClassMember) => boolean = () => true,
 ): SdsClassMember[] {
     return node?.body?.members?.filter(filterFunction) ?? [];
 };
-
 export const enumVariantsOrEmpty = function (node: SdsEnum | undefined): SdsEnumVariant[] {
     return node?.body?.variants ?? [];
 };
-
 export const importsOrEmpty = function (node: SdsModule | undefined): SdsImport[] {
     return node?.imports ?? [];
 };
-
 export const moduleMembersOrEmpty = function (node: SdsModule | undefined): SdsModuleMember[] {
     return node?.members?.filter(isSdsModuleMember) ?? [];
 };
-
 export const packageNameOrUndefined = function (node: AstNode | undefined): string | undefined {
     return getContainerOfType(node, isSdsModule)?.name;
 };
-
 export const parametersOrEmpty = function (node: SdsCallable | undefined): SdsParameter[] {
     return node?.parameterList?.parameters ?? [];
 };
-
 export const placeholdersOrEmpty = function (node: SdsBlock | undefined): SdsPlaceholder[] {
     return stream(statementsOrEmpty(node))
         .filter(isSdsAssignment)
@@ -110,19 +137,15 @@ export const placeholdersOrEmpty = function (node: SdsBlock | undefined): SdsPla
         .filter(isSdsPlaceholder)
         .toArray();
 };
-
 export const resultsOrEmpty = function (node: SdsResultList | undefined): SdsResult[] {
     return node?.results ?? [];
 };
-
 export const statementsOrEmpty = function (node: SdsBlock | undefined): SdsStatement[] {
     return node?.statements ?? [];
 };
-
 export const typeArgumentsOrEmpty = function (node: SdsTypeArgumentList | undefined): SdsTypeArgument[] {
     return node?.typeArguments ?? [];
 };
-
 export const typeParametersOrEmpty = function (node: SdsTypeParameterList | undefined): SdsTypeParameter[] {
     return node?.typeParameters ?? [];
 };
