+++ src/language/scoping/safe-ds-scope-provider.ts
@@ -47,20 +47,18 @@ import {
     SdsTypeArgument,
     SdsYield,
 } from '../generated/ast.js';
+import { isContainedIn } from '../helpers/astUtils.js';
 import {
     assigneesOrEmpty,
     classMembersOrEmpty,
-    enumVariantsOrEmpty,
-    importedDeclarationsOrEmpty,
+    enumVariantsOrEmpty, importedDeclarationsOrEmpty,
     importsOrEmpty,
-    packageNameOrUndefined,
+    isStatic, packageNameOrUndefined,
     parametersOrEmpty,
     resultsOrEmpty,
     statementsOrEmpty,
-    typeParametersOrEmpty,
-} from '../helpers/shortcuts.js';
-import { isContainedIn } from '../helpers/ast.js';
-import { isStatic } from '../helpers/checks.js';
+    typeParametersOrEmpty
+} from '../helpers/nodeProperties.js';
 import { SafeDsServices } from '../safe-ds-module.js';
 import { SafeDsTypeComputer } from '../typing/safe-ds-type-computer.js';
 import { SafeDsPackageManager } from '../workspace/safe-ds-package-manager.js';
